Creating SugarBean Resource
